+++ CHANGELOG.md
@@ -2,6 +2,11 @@
 
 ## Unreleased
 
+## v0.2.0 - 2024/06/06
+
+- Added: `Print` struct. `Output` is now deprecated, use `Print` instead (https://github.com/schneems/bullet_stream/pull/5)
+- Fix: Missing `must_use` attributes (https://github.com/schneems/bullet_stream/pull/5)
+
 ## v0.1.1 - 2024/06/03
 
 - Fix double newlines for headers (https://github.com/schneems/bullet_stream/pull/2)